Move methods out of ManagerMixin
