+++ app/models/manageiq/providers/ibm_cloud/power_virtual_servers/cloud_manager.rb
@@ -7,8 +7,6 @@ class ManageIQ::Providers::IbmCloud::PowerVirtualServers::CloudManager < ManageI
   require_nested :Template
   require_nested :Vm
 
-  include ManageIQ::Providers::IbmCloud::PowerVirtualServers::ManagerMixin
-
   has_one :network_manager,
           :foreign_key => :parent_ems_id,
           :class_name  => "ManageIQ::Providers::IbmCloud::PowerVirtualServers::NetworkManager",
@@ -50,6 +48,14 @@ class ManageIQ::Providers::IbmCloud::PowerVirtualServers::CloudManager < ManageI
     "ibm"
   end
 
+  def required_credential_fields(_type)
+    [:auth_key]
+  end
+
+  def supported_auth_attributes
+    %w[auth_key]
+  end
+
   def self.hostname_required?
     # TODO: ExtManagementSystem is validating this
     false
@@ -61,6 +67,62 @@ class ManageIQ::Providers::IbmCloud::PowerVirtualServers::CloudManager < ManageI
 
   def self.description
     @description ||= "IBM Power Systems Virtual Servers".freeze
+  end
+
+  def self.params_for_create
+    @params_for_create ||= {
+      :fields => [
+        {
+          :component  => "text-field",
+          :name       => "uid_ems",
+          :id         => "uid_ems",
+          :label      => _("PowerVS Service GUID"),
+          :isRequired => true,
+          :validate   => [{:type => "required"}],
+        },
+        {
+          :component  => "password-field",
+          :name       => "authentications.default.auth_key",
+          :id         => "authentications.default.auth_key",
+          :label      => _("IBM Cloud API Key"),
+          :type       => "password",
+          :isRequired => true,
+          :validate   => [{:type => "required"}]
+        },
+      ],
+    }.freeze
+  end
+
+  # Verify Credentials
+  # args:
+  # {
+  #   "uid_ems"         => "",
+  #   "authentications" => {
+  #     "default" => {
+  #       "auth_key" => "",
+  #     }
+  #   }
+  # }
+  def self.verify_credentials(args)
+    pcloud_guid = args["uid_ems"]
+    auth_key = args.dig("authentications", "default", "auth_key")
+    auth_key = MiqPassword.try_decrypt(auth_key)
+    auth_key ||= find(args["id"]).authentication_token('default')
+
+    !!raw_connect(auth_key, pcloud_guid)
+  end
+
+  def self.raw_connect(api_key, pcloud_guid)
+    if api_key.blank? || pcloud_guid.blank?
+      raise MiqException::MiqInvalidCredentialsError, _("Missing credentials")
+    end
+
+    require "ibm-cloud-sdk"
+    iam = IBM::Cloud::SDK::IAM.new(api_key)
+    token = iam.get_identity_token
+    power_iaas_service = IBM::Cloud::SDK::ResourceController.new(token).get_resource(pcloud_guid)
+
+    {:token => token, :guid => pcloud_guid, :crn => power_iaas_service.crn, :region => power_iaas_service.region_id, :tenant => power_iaas_service.account_id}
   end
 
   def self.create_from_params(params, endpoints, authentications)
@@ -88,6 +150,25 @@ class ManageIQ::Providers::IbmCloud::PowerVirtualServers::CloudManager < ManageI
         ems.save!
       end
     end
+  end
+
+  def connect(options = {})
+    auth_key = authentication_key(options[:auth_type])
+    creds = self.class.raw_connect(auth_key, uid_ems)
+
+    options[:service] ||= "PowerIaas"
+    case options[:service]
+    when "PowerIaas"
+      region, guid, token, crn, tenant = creds.values_at(:region, :guid, :token, :crn, :tenant)
+      IBM::Cloud::SDK::PowerIaas.new(region, guid, token, crn, tenant)
+    else
+      raise ArgumentError, "Unknown target API set: '#{options[:service]}''"
+    end
+  end
+
+  def verify_credentials(_auth_type = nil, options = {})
+    connect(options)
+    true
   end
 
   def ensure_managers
