+++ app/models/manageiq/providers/ibm_cloud/power_virtual_servers/storage_manager.rb
@@ -2,7 +2,6 @@ class ManageIQ::Providers::IbmCloud::PowerVirtualServers::StorageManager < Manag
   require_nested :CloudVolume
   require_nested :Refresher
 
-  include ManageIQ::Providers::IbmCloud::PowerVirtualServers::ManagerMixin
   include ManageIQ::Providers::StorageManager::BlockMixin
 
   delegate :authentication_check,
